+++ azure-cli-diff-tool/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+0.0.3
+++++++
+* Add prop value for parameter add/remove property
+* Remove type add break justification
+
 0.0.2
 ++++++
 * Change time consuming into info log level